// src/models/Keywords.ts
export interface Keywords {
  id?: number;
  name?: string;
  response?: string;
  triggers?: string;
  active?: number;
  ignore_cd?: number;
  command_mode?: number;
  user_level?: number;
  count?: number;
  channel_id?: number;
  creator_id?: number;
  creator_name?: string;
  channel_name?: string;
  uid?: number;
  mods_edit?: number;
  cooldown?: number;
  used_at?: number;
  created_at?: number;
}