// src/models/Timers.ts
export interface Timers {
  id?: number;
  timer_name?: string;
  timer_interval?: number;
  timer_chat_lines?: number;
  timer_timestamp?: number;
  timer_last_timer?: number;
  active?: number;
  uid?: number;
  channel_name?: string;
  twitch_id?: number;
  creator_id?: number;
  mods_edit?: number;
  count?: number;
  timer_list?: string;
  created_at?: number;
}